Adding G_control$structure = "homogeneous" which is used specifically when a corresponding custom Sigma is supplied. This allows G to a (rank-reduced) unstructured covariance matrix but must have the same values on the diagonals.
Warnings are now put it to check that if custom Sigma is not supplied, the G_control$structure could only be "unstructured".

Note the output is set up such that for both G_control$structure = "identity" and "homogenous", Sigma and G itself is scaled by the estimated constant. While not formally documented in the help file, this is fine [lambda*G \otimes Sigma = G \otimes lambda*Sigma] and is done so in prep for multiple custom_Sigma to be permitted.